feat: Use BooksAPI and router, componentize app
All of the app was contained within one file (App.js) so it was
refactored into components to make the code clearer, better structured
and more maintainable.

The book list was hard coded into the app so it was changed to use the
BooksAPI which loads the book data from a server.

The app was using the state object to implement its own routing so
this was changed to use the react router.