feat: search and update books on bookshelf
Allow searching for books and adding them to a bookshelf. Make sure
that the books in the search results have the same state as those on
the shelf.

Compononetize BooksGrid so that it can be reused in both the BookShelf
and the SearchBooks components.

Check thumbnail exists to prevent error for Books without images.

Updated the README.md to include only relevant information about
installing and starting the application.